+++ CHANGELOG.md
@@ -7,11 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2024-07-04
+
+### Added
+
+- HStack component
+
+### Changed
+
+- Update key in VStack component
+- Remove padding in input
+
 ## [1.3.0] - 2024-07-03
 
 ### Added
 
-- VStack Component
+- VStack component
 
 ## [1.2.1] - 2024-07-03